dev: move shop screen location
