+++ iwfpapp/test/shop_screen_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:iwfpapp/shop.dart';
+import 'package:iwfpapp/screens/shop/main.dart';
 
 void main() {
   testWidgets('test shop widget render no crash', (WidgetTester tester) async {
